hf_space/requirements.txt
flask>=3.1
gunicorn>=23.0
numpy>=1.24
pandas>=1.5
scikit-learn>=1.3
rdkit>=2022.9
xgboost>=2.0
catboost>=1.2
torch>=2.0
pillow>=10.0
tabulate>=0.9
google-genai>=1.14.0
